+++ changelog.md
@@ -1,3 +1,8 @@
+## 0.11.100.0
+
+ - new `hmac` and `hmaclazy` functions providing HMAC-MD5
+   computation conforming to RFC2104 and RFC2202
+
 ## 0.11.7.2
 
  - switch to 'safe' FFI for calls where overhead becomes neglible